Move ServiceFactory to Controller/Action
