+++ library/GeometriaLab/Api/Mvc/Controller/Action/Params/Params.php
@@ -7,7 +7,7 @@
  * To change this template use File | Settings | File Templates.
  */
 
-namespace GeometriaLab\Api\Mvc\Controller\Action;
+namespace GeometriaLab\Api\Mvc\Controller\Action\Params;
 
 /**
  *
